third_party/blink/web_tests/external/wpt/css/cssom/font-variant-shorthand-serialization-expected.txt

This is a testharness.js-based test.
FAIL font-variant: normal serialization assert_equals: font-variant-position was set expected "normal" but got ""
FAIL font-variant: none serialization assert_equals: font-variant-position was set expected "normal" but got ""
FAIL font-variant-ligatures: none serialization with non-default value for another longhand assert_equals: font-variant-position was set expected "normal" but got ""
FAIL font-variant: normal with non-default longhands assert_equals: font-variant-position was set expected "normal" but got ""
FAIL CSS-wide keyword in one longhand assert_equals: font-variant-position was set expected "normal" but got ""
FAIL CSS-wide keyword in shorthand assert_equals: font-variant-position was set expected "initial" but got ""
FAIL font: menu serialization assert_equals: font-variant was set expected "" but got "normal"
Harness: the test ran to completion.
